requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
SQLAlchemy-ImageAttach==1.1.0
Flask-Admin==1.5.7
